Moving DruidFilterBuilders to druid packages
