+++ CHANGELOG.md
@@ -58,6 +58,9 @@ Current
    * Made static methods into instance methods, with a default instance
    * Created a factory interface
 
+- [Refactored DruidFilterBuilders]()  
+   * Moved DruidFilterBuilder and clients to new package
+
 
 ### Deprecated:
