+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/builders/DruidOrFilterBuilder.java
@@ -1,6 +1,6 @@
-// Copyright 2016 Yahoo Inc.
+// Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.data.filterbuilders;
+package com.yahoo.bard.webservice.druid.model.builders;
 
 import com.yahoo.bard.webservice.data.dimension.Dimension;
 import com.yahoo.bard.webservice.data.dimension.DimensionRowNotFoundException;
