+++ fili-core/src/main/java/com/yahoo/bard/webservice/exception/TooManyDruidFiltersException.java
@@ -3,6 +3,7 @@
 package com.yahoo.bard.webservice.exception;
 
 import com.yahoo.bard.webservice.data.dimension.FilterBuilderException;
+import com.yahoo.bard.webservice.druid.model.builders.DruidOrFilterBuilder;
 
 /**
  * Unchecked exception for situations when too many Druid filters a generated for a Druid query.
@@ -10,7 +11,7 @@ import com.yahoo.bard.webservice.data.dimension.FilterBuilderException;
  * Dimensions with extremely large cardinalities could result in such error when user put a "contain" filter on the
  * dimension value. For example, a filter of dimension|id-contains[123], where there are 10,000 ID's starting with 123,
  * could generate 10,000 Druid filters using
- * {@link com.yahoo.bard.webservice.data.filterbuilders.DruidOrFilterBuilder}. This giant query shall eventually
+ * {@link DruidOrFilterBuilder}. This giant query shall eventually
  * timeout the Druid query and returns the timeout error to API user.
  */
 public class TooManyDruidFiltersException extends FilterBuilderException {
