+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/apirequest/utils/TestingDataApiRequestImpl.groovy
@@ -2,7 +2,7 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web.apirequest.utils
 
-import com.yahoo.bard.webservice.data.filterbuilders.DruidOrFilterBuilder
+import com.yahoo.bard.webservice.druid.model.builders.DruidOrFilterBuilder
 import com.yahoo.bard.webservice.data.time.DefaultTimeGrain
 import com.yahoo.bard.webservice.druid.model.having.Having
 import com.yahoo.bard.webservice.table.LogicalTable
